backport: [GITIAN] Pull gitian sources in our repo
This is a backport of ABC: https://reviews.bitcoinabc.org/D6972

The motivation for it is that we will need future backports that
customize the operation of gitian-builder.

---

Summary:
This will help adding features in before they are merged upstream, and
simplify setting up the builds.
Docs will be updated in follow ups diffs in order to not make review
difficult.

Notes:
 - A notice has been  added to the doc/README.md file to mention the license of `gitian-builder`.
 - The `.arclint` file is updated to prevent linter failures on the `contrib/gitian-builder` directory.
 - There are missing files (namely the `target-bin` directory. This is due to an upstream bug in the `.gitignore` file which will be fixed in a follow-up diff.