[ci] Hotfix for 'pages' job: pin mkdocs-material to 8.2.16
This cleans up the dependency installation problem which
resulted in 'invalid command bdist_wheel', as well as the
"'dict object' has no attribute 'hide'" error thrown from
within the mkdocs-material plugin for mkdocs.

The first (bdist_wheel) issue had been there already but
was not failing the job. The job failure was due to an
update of the external mkdocs-material package which
migrated from 8.2.16 to 8.3.0 .

This hot fix pins the material plugin to 8.2.16 which
was working.

Searching on the Internet shows a similar problem happened
in the old 7.x version of 'material', which some users
had to work around in a similar way.

Test plan:
- check that 'master' branch CI completes successfully,
  including the 'pages' job, after merge of this MR.

Closes #430