Code quality: Use global constant `BLOCK_HEADER_SIZE` rather than int literal `80`
Summary
-------

In `validation.cpp`, some code that reads/writes blocks to disk was using a literal
`80` in the code to represent the size of a block header.  It is preferable to
use a constant for such magic numbers.

While the block header will always be 80 for the foreseeable future -- should we
need to hard fork the chain someday to support larger headers, it would be
ideal to not have latent bugs in the code due to using hard-coded `80` and
neglecting to easily find this number in the code.

Test Plan
---------

- `ninja all check`

This is a pure code quality change -- despite touching block.h and validation.cpp,
no behavioral changes are being introduced to the codebase.

Closes #435.