[fuzz] Avoid double and triple copying of data when running fuzzers
Summary
---

Existing code was double and/or triple copying of the data coming in
from the fuzzer library, because it was passing around `std::vector`
by value as well as using `CDataStream` which itself copies the input
vector into its own internal vector.

We modified the code to avoid doing this, by utilizing our own
C++20-std::span-workalike, `Span`, and using the zero-copy
`GenericVectorReader` in place of `CDataStream` for reading the
passed-in fuzzer memory buffer.

In this way we never copy the input memory that libfuzzer provides and
the fuzzer should execute much more quickly.