Fix bug where `getrawtransaction` with verbose=2 on a coinbase txn fails
Summary
---

Fixes issue #461.  The issue was that coinbase txns have no previous
input.  But this code didn't guard against that possibility.

The options here were to set the fee to "0" or to do what Core does
in this case: don't add the "fee" key to the tx JSON.

We opted to do what Core does for compatibility and for simplicity.

Also updated the `rpc_rawtransaction.py` functional test to test for
this case.

Also in this commit: removed an un-needed .reserve() call.

Test Plan
---

- `ninja all check-all`
- See about reproducing #461 against master and test that against this
  commit it's indeed fixed and doesn't error-out on coinbase txns.