RPC: Read block data directly from disk in `getblock verbosity=0`
Summary
---

This is a performance optimization whereby we don't do any validation on
block data for `getblock` with `verbosity=0`, since doing that involves a bunch
of round-trips to deserialize and then re-serialize the block at least
twice.  For very large blocks this may be slow.

Instead, we leverage the new `ReadRawBlockFromDisk()` call merged in !1609
and use it for the `getblock` RPC in the `verbosity=0` case to just return the hex
bytes from the block as they came in off the disk.

Note that if the user is running with `-checkblockreads=1`, then
consistency checks are still done.

**Caveats to this approach:**

It's possible that a bitcoind with corrupted block data will serve up corrupt
blocks to its RPC clients. The hope is that the clients reading the block data
will detect any unlikely corruption and react accordingly.

The tradeoff is that before we would have returned an RPC error in the
unlikely case of filesystem corruption, whereas now we do not always
return errors and instead we rely on the client to validate the block data.

**Advantages to this appraoch:**

A lot less CPU time is wasted serving up blocks to block explorers that take raw
data or to services like Fulcrum.  I did a full Fulcrum resynch of testnet3 as a
test and with this MR it takes ~1350 seconds on average (versus ~1460 seconds without
this MR for a savings of 6.8% for this test).

Also note that memory fragmentation is decreased with this approach as
well (bitcoind was using less memory as I did the synch).

Test Plan
---

- `ninja all check-all`