Don't log on success by default in ReadRawBlockFromDisk with `-checkblockreads=1`
Summary
---

To reduce log spam, we turn an extra `LogPrintf()` into a `LogPrint(BCLog::BENCH, ...)`
when the extra checks requested by `-checkblockreads=1` yield a successfully-read
block (this is the common case 99.9999999% of the time).

Unconditionally printing to the log in this case will just yield 700k extra log lines
on a full resynch with e.g. Fulcrum and is very wasteful and unnecessary.  Instead
we make it a "bench" log category so that it's only printed in the benchmark case.
We also added an additional "Tic" timing to the log line to tell the user how long it
took to do the check.

Test Plan
---

- `ninja all check-all`