Added new functions to secp256k1 for serializing the pubkey from a multiset
Summary
-------

This is part of initial work for the UTXO Fast Sync project.

As per the spec, we will be using the multiset pubkey (rather than the
pubkey hash) to identify a UTXO set (or UTXO commitment).
Using the pubkey has the advantage that it means we can always reconstruct
the multiset object given a serialized pubkey.

The code for ec-multiset we have in secp256k1 could only produce a hash
of the ec-multiset, and not the pubkey itself.  However, the pubkey is kept
in memory internally.  This MR adds new functions to the secp256k1 lib to
retrieve the pubkey:

- Serialize an empty or non-empty multiset to a 33 byte compressed pubkey.
- Unserialize a 33-byte compressed pubkey to an ec multiset.
- Test if an ec-multiset is empty.

This MR also adds a bunch of tests for the above.

This MR also enables the EC multiset module in secp256k1 because it will
be needed for subsequent MRs that will depend on this one.
In case you are wondering, it's safe to enable it, since if you don't use
it, it does no harm and is inert in the codebase.

I checked with Core and ABC and no subsequent work or commits exist for
the ec-multiset module, which was originally contributed to ABC here:

https://reviews.bitcoinabc.org/D1072

Core has an extant PR for similar work here which was never merged:

https://github.com/bitcoin-core/secp256k1/pull/477.

With this MR we will have the most feature-rich implementation of the module.

Test Plan
---------

- `ninja check-secp256k1`

(you may have to clear your cmake cache and/or start from a clean build/
directory because of bugs in our cmake building for secp256k1)